+++ lib/gptk.rb
@@ -11,7 +11,7 @@ end
 # GPT Kit - A collection of useful tools for interacting with GPT agents and generating content
 module GPTK
   START_TIME = Time.now
-  VERSION = '0.15'.freeze
+  VERSION = '0.16'.freeze
 
   # Load configuration files
   Config.load_openai_setup